@@ -18,19 +18,19 @@
   },
   "homepage": "https://github.com/eivers88/responsive-css-sprite-generator#readme",
   "devDependencies": {
-    "babel-core": "^6.24.1",
+    "babel-core": "^6.25.0",
     "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.24.1",
-    "browser-sync": "^2.18.8",
+    "browser-sync": "^2.18.12",
     "concurrently": "^3.4.0",
     "expose-loader": "^0.7.3",
     "gulp": "^3.9.1",
     "gulp-autoprefixer": "^3.1.1",
     "gulp-ruby-sass": "^2.1.1",
     "gulp-sourcemaps": "^2.6.0",
-    "webpack": "^2.4.1"
+    "webpack": "^2.6.1"
   },
   "dependencies": {
-    "clipboard": "^1.6.1"
+    "clipboard": "^1.7.1"
   }
 }
